/* PR_START p/alias_same_dimension 02 */ Move alias rendering tests to a separate file
