+++ tests_metricflow/snapshots/test_alias_rendering.py/SqlPlan/DuckDB/test_simple_metric_constraint_with_single_expr_and_alias__plan0.sql
@@ -1,6 +1,6 @@
 test_name: test_simple_metric_constraint_with_single_expr_and_alias
-test_filename: test_query_rendering.py
-sql_engine: Databricks
+test_filename: test_alias_rendering.py
+sql_engine: DuckDB
 ---
 -- Write to DataTable
 SELECT
@@ -184,7 +184,7 @@ FROM (
                   , EXTRACT(quarter FROM bookings_source_src_28000.ds) AS ds__extract_quarter
                   , EXTRACT(month FROM bookings_source_src_28000.ds) AS ds__extract_month
                   , EXTRACT(day FROM bookings_source_src_28000.ds) AS ds__extract_day
-                  , EXTRACT(DAYOFWEEK_ISO FROM bookings_source_src_28000.ds) AS ds__extract_dow
+                  , EXTRACT(isodow FROM bookings_source_src_28000.ds) AS ds__extract_dow
                   , EXTRACT(doy FROM bookings_source_src_28000.ds) AS ds__extract_doy
                   , DATE_TRUNC('day', bookings_source_src_28000.ds_partitioned) AS ds_partitioned__day
                   , DATE_TRUNC('week', bookings_source_src_28000.ds_partitioned) AS ds_partitioned__week
@@ -195,7 +195,7 @@ FROM (
                   , EXTRACT(quarter FROM bookings_source_src_28000.ds_partitioned) AS ds_partitioned__extract_quarter
                   , EXTRACT(month FROM bookings_source_src_28000.ds_partitioned) AS ds_partitioned__extract_month
                   , EXTRACT(day FROM bookings_source_src_28000.ds_partitioned) AS ds_partitioned__extract_day
-                  , EXTRACT(DAYOFWEEK_ISO FROM bookings_source_src_28000.ds_partitioned) AS ds_partitioned__extract_dow
+                  , EXTRACT(isodow FROM bookings_source_src_28000.ds_partitioned) AS ds_partitioned__extract_dow
                   , EXTRACT(doy FROM bookings_source_src_28000.ds_partitioned) AS ds_partitioned__extract_doy
                   , DATE_TRUNC('day', bookings_source_src_28000.paid_at) AS paid_at__day
                   , DATE_TRUNC('week', bookings_source_src_28000.paid_at) AS paid_at__week
@@ -206,7 +206,7 @@ FROM (
                   , EXTRACT(quarter FROM bookings_source_src_28000.paid_at) AS paid_at__extract_quarter
                   , EXTRACT(month FROM bookings_source_src_28000.paid_at) AS paid_at__extract_month
                   , EXTRACT(day FROM bookings_source_src_28000.paid_at) AS paid_at__extract_day
-                  , EXTRACT(DAYOFWEEK_ISO FROM bookings_source_src_28000.paid_at) AS paid_at__extract_dow
+                  , EXTRACT(isodow FROM bookings_source_src_28000.paid_at) AS paid_at__extract_dow
                   , EXTRACT(doy FROM bookings_source_src_28000.paid_at) AS paid_at__extract_doy
                   , bookings_source_src_28000.is_instant AS booking__is_instant
                   , DATE_TRUNC('day', bookings_source_src_28000.ds) AS booking__ds__day
@@ -218,7 +218,7 @@ FROM (
                   , EXTRACT(quarter FROM bookings_source_src_28000.ds) AS booking__ds__extract_quarter
                   , EXTRACT(month FROM bookings_source_src_28000.ds) AS booking__ds__extract_month
                   , EXTRACT(day FROM bookings_source_src_28000.ds) AS booking__ds__extract_day
-                  , EXTRACT(DAYOFWEEK_ISO FROM bookings_source_src_28000.ds) AS booking__ds__extract_dow
+                  , EXTRACT(isodow FROM bookings_source_src_28000.ds) AS booking__ds__extract_dow
                   , EXTRACT(doy FROM bookings_source_src_28000.ds) AS booking__ds__extract_doy
                   , DATE_TRUNC('day', bookings_source_src_28000.ds_partitioned) AS booking__ds_partitioned__day
                   , DATE_TRUNC('week', bookings_source_src_28000.ds_partitioned) AS booking__ds_partitioned__week
@@ -229,7 +229,7 @@ FROM (
                   , EXTRACT(quarter FROM bookings_source_src_28000.ds_partitioned) AS booking__ds_partitioned__extract_quarter
                   , EXTRACT(month FROM bookings_source_src_28000.ds_partitioned) AS booking__ds_partitioned__extract_month
                   , EXTRACT(day FROM bookings_source_src_28000.ds_partitioned) AS booking__ds_partitioned__extract_day
-                  , EXTRACT(DAYOFWEEK_ISO FROM bookings_source_src_28000.ds_partitioned) AS booking__ds_partitioned__extract_dow
+                  , EXTRACT(isodow FROM bookings_source_src_28000.ds_partitioned) AS booking__ds_partitioned__extract_dow
                   , EXTRACT(doy FROM bookings_source_src_28000.ds_partitioned) AS booking__ds_partitioned__extract_doy
                   , DATE_TRUNC('day', bookings_source_src_28000.paid_at) AS booking__paid_at__day
                   , DATE_TRUNC('week', bookings_source_src_28000.paid_at) AS booking__paid_at__week
@@ -240,7 +240,7 @@ FROM (
                   , EXTRACT(quarter FROM bookings_source_src_28000.paid_at) AS booking__paid_at__extract_quarter
                   , EXTRACT(month FROM bookings_source_src_28000.paid_at) AS booking__paid_at__extract_month
                   , EXTRACT(day FROM bookings_source_src_28000.paid_at) AS booking__paid_at__extract_day
-                  , EXTRACT(DAYOFWEEK_ISO FROM bookings_source_src_28000.paid_at) AS booking__paid_at__extract_dow
+                  , EXTRACT(isodow FROM bookings_source_src_28000.paid_at) AS booking__paid_at__extract_dow
                   , EXTRACT(doy FROM bookings_source_src_28000.paid_at) AS booking__paid_at__extract_doy
                   , bookings_source_src_28000.listing_id AS listing
                   , bookings_source_src_28000.guest_id AS guest
