--- tests_metricflow/snapshots/test_query_rendering.py/SqlPlan/Databricks/test_simple_metric_constraint_with_single_expr_and_alias__plan0_optimized.sql
@@ -1,31 +1,0 @@
-test_name: test_simple_metric_constraint_with_single_expr_and_alias
-test_filename: test_query_rendering.py
-sql_engine: Databricks
----
--- Compute Metrics via Expressions
--- Write to DataTable
-SELECT
-  metric_time__day
-  , delayed_bookings * 2 AS double_counted_delayed_bookings
-FROM (
-  -- Constrain Output with WHERE
-  -- Select: ['__bookings', 'metric_time__day']
-  -- Aggregate Inputs for Simple Metrics
-  -- Compute Metrics via Expressions
-  SELECT
-    metric_time__day
-    , SUM(bookings) AS delayed_bookings
-  FROM (
-    -- Read Elements From Semantic Model 'bookings_source'
-    -- Metric Time Dimension 'ds'
-    -- Select: ['__bookings', 'booking__is_instant', 'metric_time__day']
-    SELECT
-      DATE_TRUNC('day', ds) AS metric_time__day
-      , is_instant AS booking__is_instant
-      , 1 AS bookings
-    FROM ***************************.fct_bookings bookings_source_src_28000
-  ) subq_10
-  WHERE NOT booking__is_instant
-  GROUP BY
-    metric_time__day
-) subq_14
